v0.07.11 - move BasicRatio base class to ratiosClasses folder
