+++ invKeepBackend/ratios/ratiosClasses/PBRatio.js
@@ -1,5 +1,5 @@
 const SharedJS = require('../../../invKeepFrontend/src/app/shared/sharedJS');
-const BasicRatio = require('../BasicRatio')
+const BasicRatio = require('./BasicRatio')
 
 module.exports = class PBRatio extends BasicRatio {
